feat: implement shadcn global theme and moved components to shadcn generated folder
