+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/public/C.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/C.png">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Chinjen.</title>
   </head>
