--- src/app/portfolio/components/DarkModeToggle.tsx
@@ -1,33 +1,0 @@
-import { useState, useEffect } from 'react';
-
-const DarkModeToggle: React.FC = () => {
-  const [isDarkMode, setIsDarkMode] = useState(() => {
-    // Check localStorage for saved preference
-    const savedMode = localStorage.getItem('theme');
-    return savedMode ? savedMode === 'dark' : false;
-  });
-
-  useEffect(() => {
-    // Apply the dark mode class to the root element based on state
-    const root = document.documentElement;
-    if (isDarkMode) {
-      root.classList.add('dark');
-      localStorage.setItem('theme', 'dark');
-    } else {
-      root.classList.remove('dark');
-      localStorage.setItem('theme', 'light');
-    }
-  }, [isDarkMode]);
-
-  const toggleDarkMode = () => {
-    setIsDarkMode(!isDarkMode);
-  };
-
-  return (
-    <button onClick={toggleDarkMode} className="rounded bg-gray-200 p-2 dark:bg-gray-800">
-      {isDarkMode ? 'Light Mode' : 'Dark Mode'}
-    </button>
-  );
-};
-
-export default DarkModeToggle;
