+++ src/components/mode-toggle.tsx
@@ -1,6 +1,7 @@
-import { Moon, Sun } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { useTheme } from '@/components/theme-provider';
+import { FaSun } from 'react-icons/fa';
+import { BsFillMoonStarsFill } from 'react-icons/bs';
 
 export function ModeToggle() {
   const { theme, setTheme } = useTheme();
@@ -21,9 +22,9 @@ export function ModeToggle() {
       className="border-none bg-inherit shadow-none hover:bg-inherit dark:bg-inherit hover:dark:bg-inherit"
     >
       {theme === 'dark' ? (
-        <Sun className="h-[1.2rem] w-[1.2rem] text-yellow-500 transition-all" />
+        <FaSun className="h-[1.2rem] w-[1.2rem] text-yellow-500 transition-all" />
       ) : (
-        <Moon className="h-[1.2rem] w-[1.2rem] text-gray-800 transition-all dark:text-gray-100" />
+        <BsFillMoonStarsFill className="h-[1.2rem] w-[1.2rem] text-gray-800 transition-all dark:text-gray-100" />
       )}
       <span className="sr-only">Toggle theme</span>
     </Button>
